Move DB gateway to own package
