+++ src/main/java/LoginUseCase/LoginInteractor.java
@@ -1,5 +1,6 @@
 package LoginUseCase;
 
+import dbGateway.DatabaseGateway;
 import screens.Presenter;
 
 public class LoginInteractor implements LoginInputBoundary, Interactor{
