+++ src/main/java/dbGateway/DatabaseGateway.java
@@ -1,4 +1,4 @@
-package LoginUseCase;
+package dbGateway;
 
 public interface DatabaseGateway {
     boolean validateAndLogin(String email, String password);
